requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
python-dotenv==1.1.1
PyMuPDF==1.26.3
python-docx==0.8.11
scikit-learn==1.2.2
numpy==1.25.0
pandas==2.2.2
pinecone-client>=3.1.0
requests==2.31.0
python-multipart==0.0.6
